all: gcc warning clean (round 2) * reorder static initialization list * Add missing virtual desctrutor to virtual object * int -> uint/u32/uint32 cast of for loop index * add a missing return in pxTrySetFocus * fix size parameter of memset m_clut * disable missing-field-initializers warning * disable unused-local-typedefs warning
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@5685 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ pcsx2/x86/iCore.cpp
@@ -82,7 +82,7 @@ int  _getFreeXMMreg()
 	int i, tempi;
 	u32 bestcount = 0x10000;
 
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[(i+s_xmmchecknext)%iREGCNT_XMM].inuse == 0) {
 			int ret = (s_xmmchecknext+i)%iREGCNT_XMM;
 			s_xmmchecknext = (s_xmmchecknext+i+1)%iREGCNT_XMM;
@@ -91,7 +91,7 @@ int  _getFreeXMMreg()
 	}
 
 	// check for dead regs
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].needed) continue;
 		if (xmmregs[i].type == XMMTYPE_GPRREG ) {
 			if (!(EEINST_ISLIVEXMM(xmmregs[i].reg))) {
@@ -102,7 +102,7 @@ int  _getFreeXMMreg()
 	}
 
 	// check for future xmm usage
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].needed) continue;
 		if (xmmregs[i].type == XMMTYPE_GPRREG ) {
 			if( !(g_pCurInstInfo->regs[xmmregs[i].reg] & EEINST_XMM) ) {
@@ -114,7 +114,7 @@ int  _getFreeXMMreg()
 
 	tempi = -1;
 	bestcount = 0xffff;
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].needed) continue;
 		if (xmmregs[i].type != XMMTYPE_TEMP) {
 
@@ -157,7 +157,7 @@ int _allocVFtoXMMreg(VURegs *VU, int xmmreg, int vfreg, int mode) {
 	int i;
 	int readfromreg = -1;
 
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if ((xmmregs[i].inuse == 0)  || (xmmregs[i].type != XMMTYPE_VFREG) ||
 		     (xmmregs[i].reg != vfreg) || (xmmregs[i].VU != XMM_CONV_VU(VU)))
 			continue;
@@ -211,7 +211,7 @@ int _checkXMMreg(int type, int reg, int mode)
 {
 	int i;
 
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].inuse && (xmmregs[i].type == (type&0xff)) && (xmmregs[i].reg == reg)) {
 
 			if ( !(xmmregs[i].mode & MODE_READ) ) {
@@ -240,7 +240,7 @@ int _allocACCtoXMMreg(VURegs *VU, int xmmreg, int mode) {
 	int i;
 	int readfromreg = -1;
 
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].inuse == 0) continue;
 		if (xmmregs[i].type != XMMTYPE_ACC) continue;
 		if (xmmregs[i].VU != XMM_CONV_VU(VU) ) continue;
@@ -296,7 +296,7 @@ int _allocACCtoXMMreg(VURegs *VU, int xmmreg, int mode) {
 int _allocFPtoXMMreg(int xmmreg, int fpreg, int mode) {
 	int i;
 
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].inuse == 0) continue;
 		if (xmmregs[i].type != XMMTYPE_FPREG) continue;
 		if (xmmregs[i].reg != fpreg) continue;
@@ -333,7 +333,7 @@ int _allocGPRtoXMMreg(int xmmreg, int gprreg, int mode)
 {
 	int i;
 
-	for (i=0; i<iREGCNT_XMM; i++)
+	for (i=0; (uint)i<iREGCNT_XMM; i++)
 	{
 		if (xmmregs[i].inuse == 0) continue;
 		if (xmmregs[i].type != XMMTYPE_GPRREG) continue;
@@ -439,7 +439,7 @@ int _allocFPACCtoXMMreg(int xmmreg, int mode)
 {
 	int i;
 
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].inuse == 0) continue;
 		if (xmmregs[i].type != XMMTYPE_FPACC) continue;
 
@@ -477,7 +477,7 @@ int _allocFPACCtoXMMreg(int xmmreg, int mode)
 void _addNeededVFtoXMMreg(int vfreg) {
 	int i;
 
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].inuse == 0) continue;
 		if (xmmregs[i].type != XMMTYPE_VFREG) continue;
 		if (xmmregs[i].reg != vfreg) continue;
@@ -491,7 +491,7 @@ void _addNeededGPRtoXMMreg(int gprreg)
 {
 	int i;
 
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].inuse == 0) continue;
 		if (xmmregs[i].type != XMMTYPE_GPRREG) continue;
 		if (xmmregs[i].reg != gprreg) continue;
@@ -505,7 +505,7 @@ void _addNeededGPRtoXMMreg(int gprreg)
 void _addNeededACCtoXMMreg() {
 	int i;
 
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].inuse == 0) continue;
 		if (xmmregs[i].type != XMMTYPE_ACC) continue;
 
@@ -518,7 +518,7 @@ void _addNeededACCtoXMMreg() {
 void _addNeededFPtoXMMreg(int fpreg) {
 	int i;
 
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].inuse == 0) continue;
 		if (xmmregs[i].type != XMMTYPE_FPREG) continue;
 		if (xmmregs[i].reg != fpreg) continue;
@@ -532,7 +532,7 @@ void _addNeededFPtoXMMreg(int fpreg) {
 void _addNeededFPACCtoXMMreg() {
 	int i;
 
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].inuse == 0) continue;
 		if (xmmregs[i].type != XMMTYPE_FPACC) continue;
 
@@ -545,7 +545,7 @@ void _addNeededFPACCtoXMMreg() {
 void _clearNeededXMMregs() {
 	int i;
 
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 
 		if( xmmregs[i].needed ) {
 
@@ -566,7 +566,7 @@ void _deleteVFtoXMMreg(int reg, int vu, int flush)
 	int i;
 	VURegs *VU = vu ? &VU1 : &VU0;
 
-	for (i=0; i<iREGCNT_XMM; i++)
+	for (i=0; (uint)i<iREGCNT_XMM; i++)
 	{
 		if (xmmregs[i].inuse && (xmmregs[i].type == XMMTYPE_VFREG) &&
 		   (xmmregs[i].reg == reg) && (xmmregs[i].VU == vu))
@@ -585,7 +585,7 @@ void _deleteVFtoXMMreg(int reg, int vu, int flush)
 							if( xmmregs[i].mode & MODE_VUZ )
 							{
 								// xyz, don't destroy w
-								int t0reg;
+								uint t0reg;
 
 								for (t0reg = 0; t0reg < iREGCNT_XMM; ++t0reg)
 								{
@@ -637,7 +637,7 @@ void _deleteACCtoXMMreg(int vu, int flush)
 	int i;
 	VURegs *VU = vu ? &VU1 : &VU0;
 
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].inuse && (xmmregs[i].type == XMMTYPE_ACC) && (xmmregs[i].VU == vu)) {
 
 			switch(flush) {
@@ -652,7 +652,7 @@ void _deleteACCtoXMMreg(int vu, int flush)
 
 							if( xmmregs[i].mode & MODE_VUZ ) {
 								// xyz, don't destroy w
-								int t0reg;
+								uint t0reg;
 								for(t0reg = 0; t0reg < iREGCNT_XMM; ++t0reg ) {
 									if( !xmmregs[t0reg].inuse ) break;
 								}
@@ -697,7 +697,7 @@ void _deleteACCtoXMMreg(int vu, int flush)
 void _deleteGPRtoXMMreg(int reg, int flush)
 {
 	int i;
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 
 		if (xmmregs[i].inuse && xmmregs[i].type == XMMTYPE_GPRREG && xmmregs[i].reg == reg ) {
 
@@ -731,7 +731,7 @@ void _deleteGPRtoXMMreg(int reg, int flush)
 void _deleteFPtoXMMreg(int reg, int flush)
 {
 	int i;
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].inuse && xmmregs[i].type == XMMTYPE_FPREG && xmmregs[i].reg == reg ) {
 			switch(flush) {
 				case 0:
@@ -771,7 +771,7 @@ void _freeXMMreg(int xmmreg)
 				if( xmmregs[xmmreg].mode & MODE_VUZ )
 				{
 					// don't destroy w
-					int t0reg;
+					uint t0reg;
 					for(t0reg = 0; t0reg < iREGCNT_XMM; ++t0reg ) {
 						if( !xmmregs[t0reg].inuse ) break;
 					}
@@ -812,7 +812,7 @@ void _freeXMMreg(int xmmreg)
 				if( xmmregs[xmmreg].mode & MODE_VUZ )
 				{
 					// don't destroy w
-					int t0reg;
+					uint t0reg;
 
 					for(t0reg = 0; t0reg < iREGCNT_XMM; ++t0reg ) {
 						if( !xmmregs[t0reg].inuse ) break;
@@ -871,7 +871,7 @@ void _freeXMMreg(int xmmreg)
 int _getNumXMMwrite()
 {
 	int num = 0, i;
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if( xmmregs[i].inuse && (xmmregs[i].mode&MODE_WRITE) ) ++num;
 	}
 
@@ -882,12 +882,12 @@ u8 _hasFreeXMMreg()
 {
 	int i;
 
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (!xmmregs[i].inuse) return 1;
 	}
 
 	// check for dead regs
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].needed) continue;
 		if (xmmregs[i].type == XMMTYPE_GPRREG ) {
 			if( !EEINST_ISLIVEXMM(xmmregs[i].reg) ) {
@@ -897,7 +897,7 @@ u8 _hasFreeXMMreg()
 	}
 
 	// check for dead regs
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].needed) continue;
 		if (xmmregs[i].type == XMMTYPE_GPRREG  ) {
 			if( !(g_pCurInstInfo->regs[xmmregs[i].reg]&EEINST_USED) ) {
@@ -913,7 +913,7 @@ void _moveXMMreg(int xmmreg)
 	int i;
 	if( !xmmregs[xmmreg].inuse ) return;
 
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].inuse) continue;
 		break;
 	}
@@ -933,7 +933,7 @@ void _flushXMMregs()
 {
 	int i;
 
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].inuse == 0) continue;
 
 		pxAssert( xmmregs[i].type != XMMTYPE_TEMP );
@@ -950,7 +950,7 @@ void _freeXMMregs()
 {
 	int i;
 
-	for (i=0; i<iREGCNT_XMM; i++) {
+	for (i=0; (uint)i<iREGCNT_XMM; i++) {
 		if (xmmregs[i].inuse == 0) continue;
 
 		pxAssert( xmmregs[i].type != XMMTYPE_TEMP );
